SageServe-original/requirements.txt
hydra-core==1.3.2
hydra-joblib-launcher==1.2.0
hydra-ray-launcher==1.2.1
matplotlib==3.7.2
networkx==3.1
numpy==1.25.2
omegaconf==2.3.0
pandas==2.0.3
plotly==5.17.0
ray==2.6.3
requests==2.31.0
scikit_learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
nbformat==5.9.2
pulp==2.9.0
sortedcontainers==2.4.0
tqdm==4.67.1